Embed path_resolver contents into the Elaborator
